Hoist command registry and dispatch from IDracManager to RedfishManager
Move the command registry and the invoke/sync_invoke/async_invoke dispatch
machinery from the Dell-specific IDracManager up into the product-neutral
RedfishManager parent. Commands stay IDracManager subclasses and inherit the
machinery one MRO hop higher, so behavior is unchanged.

- The shared registry becomes collections.defaultdict(dict), dropping the
  ApiRequestType dependency from the neutral layer; an unknown api_call now
  raises UnsupportedAction instead of KeyError.
- Dispatch type hints use Hashable so RedfishManager never imports iDRAC code.
- Resync the span-root and config-loader gate baselines to the shifted line
  numbers in both manager files.
- Add tests covering the shared registry, inherited dispatch, and the
  UnsupportedAction path.
