+++ tools/config_loader_baseline.txt
@@ -13,7 +13,7 @@ redfish_ctl/discovery/cmd_discovery.py:258
 redfish_ctl/discovery/cmd_discovery.py:260
 redfish_ctl/discovery/cmd_discovery.py:262
 redfish_ctl/fleet/cmd_fleet.py:53
-redfish_ctl/idrac_manager.py:527
+redfish_ctl/idrac_manager.py:352
 redfish_ctl/licenses/cmd_dell_license_actions.py:468
 redfish_ctl/licenses/cmd_dell_license_actions.py:472
 redfish_ctl/licenses/cmd_license_install.py:202
@@ -27,10 +27,10 @@ redfish_ctl/oem/cmd_nvidia_debug_token.py:317
 redfish_ctl/oem/cmd_nvidia_debug_token.py:319
 redfish_ctl/redfish_csdl.py:70
 redfish_ctl/redfish_main.py:177
-redfish_ctl/redfish_manager.py:360
-redfish_ctl/redfish_manager.py:362
-redfish_ctl/redfish_manager.py:364
-redfish_ctl/redfish_manager.py:391
+redfish_ctl/redfish_manager.py:586
+redfish_ctl/redfish_manager.py:588
+redfish_ctl/redfish_manager.py:590
+redfish_ctl/redfish_manager.py:617
 redfish_ctl/telemetry/cmd_exporter.py:812
 redfish_ctl/telemetry/cmd_exporter.py:813
 redfish_ctl/telemetry/exporter.py:1157
