+++ tools/span_root_baseline.txt
@@ -4,7 +4,7 @@
 redfish_ctl/cmd_wait.py:38
 redfish_ctl/discover/cli.py:220
 redfish_ctl/discovery/net_scan.py:61
-redfish_ctl/idrac_manager.py:484
-redfish_ctl/idrac_manager.py:492
-redfish_ctl/redfish_manager.py:326
-redfish_ctl/redfish_manager.py:334
+redfish_ctl/idrac_manager.py:309
+redfish_ctl/idrac_manager.py:317
+redfish_ctl/redfish_manager.py:552
+redfish_ctl/redfish_manager.py:560
